Rename file of LogRepository
